fix itoa

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: kinamura <kinamura@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/27 13:46:18 by kinamura          #+#    #+#             */
-/*   Updated: 2024/06/29 07:23:23 by kinamura         ###   ########.fr       */
+/*   Updated: 2024/06/29 07:24:54 by kinamura         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ static int	ft_isspace(int c)
 
 int	ft_atoi(const char *str)
 {
-	int		sign;
-	long	ln;
+	int			sign;
+	long long	ln;
 
 	sign = 1;
 	ln = 0;
